+++ app.rb
@@ -4,23 +4,13 @@ require 'sinatra'
 require 'sinatra/reloader'
 require 'json'
 require 'rack/flash'
+require './helpers/helpers'
+
+configure do
+  use Rack::Flash
+end
 
 enable :sessions
-
-helpers do
-  def h(param)
-    Rack::Utils.escape_html(param)
-  end
-
-  def import_json
-    memo_file = File.open('json/memodb.json').read
-    @memos = JSON.parse(memo_file) if memo_file != ''
-  end
-
-  def fetch_memo(memos, params)
-    @memo = memos['memos'][fetch_memo_number(memos, params)]
-  end
-end
 
 def fetch_memo_number(memos, memo_param)
   memos['memos'].find_index { |data| data['id'].to_i == memo_param['id'].to_i }
@@ -80,7 +70,6 @@ end
 patch '/memo/:id/context' do
   if params['title'] == ''
     flash[:danger] = 'タイトルが入力されていません。'
-    p params['text']
     fetch_memo(import_json, params)
     @memo['title'] = params['title']
     @memo['text'] = params['text']
